requirements.txt
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
google-auth==2.6.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
Pillow==9.0.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
